move code for C67 based DSP PCB into it's own device (nw)
