+++ src/mame/drivers/namcos21.cpp
@@ -58,8 +58,10 @@ shade pixels according to their depth.
 
 STATUS:
     Solvaou:
+		random crashes
+
     Air Combat:
-         no polys - why?
+         priority issues
 
     Winning Run
     Winning Run 91
@@ -68,9 +70,8 @@ STATUS:
           - posirq handling broken
 
     Driver's Eyes
-        Left and Right screen
+        add Left and Right screen
 
-TODO:   (*) Extract DSP BIOS
 
 TODO:   namcoic.c: in StarBlade, the sprite list is stored at a different location during startup tests.
         What register controls this?
@@ -538,10 +539,6 @@ Filter Board
 #define ENABLE_LOGGING      0
 
 
-int32_t namcos21_state::read_pointrom_data(unsigned offset)
-{
-	return m_ptrom24[offset];
-}
 
 READ16_MEMBER(namcos21_state::namcos21_video_enable_r)
 {
@@ -557,633 +554,7 @@ WRITE16_MEMBER(namcos21_state::namcos21_video_enable_w)
 	}
 }
 
-
 /***********************************************************/
-WRITE16_MEMBER(namcos21_state::dspcuskey_w)
-{ /* TODO: proper cuskey emulation */
-}
-
-READ16_MEMBER(namcos21_state::dspcuskey_r)
-{
-	uint16_t result = 0;
-	if( m_gametype == NAMCOS21_SOLVALOU )
-	{
-		switch( m_c67master->pc() )
-		{
-		case 0x805e: result = 0x0000; break;
-		case 0x805f: result = 0xfeba; break;
-		case 0x8067: result = 0xffff; break;
-		case 0x806e: result = 0x0145; break;
-		default:
-			logerror( "unk cuskey_r; pc=0x%x\n", m_c67master->pc() );
-			break;
-		}
-	}
-	else if( m_gametype == NAMCOS21_CYBERSLED )
-	{
-		switch( m_c67master->pc() )
-		{
-		case 0x8061: result = 0xfe95; break;
-		case 0x8069: result = 0xffff; break;
-		case 0x8070: result = 0x016A; break;
-		default:
-			break;
-		}
-	}
-	else if( m_gametype == NAMCOS21_AIRCOMBAT )
-	{
-		switch( m_c67master->pc() )
-		{
-		case 0x8062: result = 0xfeb9; break;
-		case 0x806a: result = 0xffff; break;
-		case 0x8071: result = 0x0146; break;
-		default:
-			break;
-		}
-	}
-	return result;
-}
-
-void namcos21_state::transmit_word_to_slave(uint16_t data)
-{
-	unsigned offs = m_mpDspState->slaveInputStart+m_mpDspState->slaveBytesAvailable++;
-	m_mpDspState->slaveInputBuffer[offs%DSP_BUF_MAX] = data;
-	if (ENABLE_LOGGING) logerror( "+%04x(#%04x)\n", data, m_mpDspState->slaveBytesAvailable );
-	m_mpDspState->slaveActive = 1;
-	if( m_mpDspState->slaveBytesAvailable >= DSP_BUF_MAX )
-	{
-		fatalerror( "IDC overflow\n" );
-	}
-}
-
-void namcos21_state::transfer_dsp_data()
-{
-	uint16_t addr = m_mpDspState->masterSourceAddr;
-	int mode = addr&0x8000;
-	addr&=0x7fff;
-	if( addr )
-	{
-		for(;;)
-		{
-			int i;
-			uint16_t old = addr;
-			uint16_t code = m_dspram16[addr++];
-			if( code == 0xffff )
-			{
-				if( mode )
-				{
-					addr = m_dspram16[addr];
-					m_mpDspState->masterSourceAddr = addr;
-					if (ENABLE_LOGGING) logerror( "LOOP:0x%04x\n", addr );
-					addr&=0x7fff;
-					if( old==addr )
-					{
-						return;
-					}
-				}
-				else
-				{
-					m_mpDspState->masterSourceAddr = 0;
-					return;
-				}
-			}
-			else if( mode==0 )
-			{ /* direct data transfer */
-				if (ENABLE_LOGGING) logerror( "DATA TFR(0x%x)\n", code );
-				transmit_word_to_slave(code);
-				for( i=0; i<code; i++ )
-				{
-					uint16_t data = m_dspram16[addr++];
-					transmit_word_to_slave(data);
-				}
-			}
-			else if( code==0x18 || code==0x1a )
-			{
-				if (ENABLE_LOGGING) logerror( "HEADER TFR(0x%x)\n", code );
-				transmit_word_to_slave(code+1);
-				for( i=0; i<code; i++ )
-				{
-					uint16_t data = m_dspram16[addr++];
-					transmit_word_to_slave(data);
-				}
-			}
-			else
-			{
-				int32_t masterAddr = read_pointrom_data(code);
-				if (ENABLE_LOGGING) logerror( "OBJ TFR(0x%x)\n", code );
-				{
-					uint16_t len = m_dspram16[addr++];
-					for(;;)
-					{
-						int subAddr = read_pointrom_data(masterAddr++);
-						if( subAddr==0xffffff )
-						{
-							break;
-						}
-						else
-						{
-							int primWords = (uint16_t)read_pointrom_data(subAddr++);
-							// TODO: this function causes an IDC overflow in Solvalou, something else failed prior to that?
-							// In Header TFR when bad parameters happens there's a suspicious 0x000f 0x0003 as first two words,
-							// maybe it's supposed to have a different length there ...
-							// cfr: object code 0x17 in service mode
-							if( primWords>2 )
-							{
-								transmit_word_to_slave(0); /* pad1 */
-								transmit_word_to_slave(len+1);
-								for( i=0; i<len; i++ )
-								{ /* transform */
-									transmit_word_to_slave(m_dspram16[addr+i]);
-								}
-								transmit_word_to_slave(0); /* pad2 */
-								transmit_word_to_slave(primWords+1);
-								for( i=0; i<primWords; i++ )
-								{
-									transmit_word_to_slave((uint16_t)read_pointrom_data(subAddr+i));
-								}
-							}
-							else
-							{
-								if (ENABLE_LOGGING) logerror( "TFR NOP?\n" );
-							}
-						}
-					} /* for(;;) */
-					addr+=len;
-				}
-			}
-		} /* for(;;) */
-	}
-}
-
-
-
-void namcos21_state::namcos21_kickstart_hacks(int internal)
-{
-	/* patch dsp watchdog */
-	switch (m_gametype)
-	{
-	case namcos21_state::NAMCOS21_AIRCOMBAT:
-		m_master_dsp_ram[0x008e] = 0x808f;
-		break;
-	case namcos21_state::NAMCOS21_SOLVALOU:
-		m_master_dsp_ram[0x008b] = 0x808c;
-		break;
-	default:
-		break;
-	}
-	if (internal)
-	{
-		if (m_mbNeedsKickstart == 0) return;
-		m_mbNeedsKickstart--;
-		if (m_mbNeedsKickstart) return;
-	}
-
-	m_namcos21_3d->swap_and_clear_poly_framebuffer();
-	m_mpDspState->masterSourceAddr = 0;
-	m_mpDspState->slaveOutputSize = 0;
-	m_mpDspState->masterFinished = 0;
-	m_mpDspState->slaveActive = 0;
-	m_c67master->set_input_line(0, HOLD_LINE);
-	m_c67slave[0]->pulse_input_line(INPUT_LINE_RESET, attotime::zero);
-}
-
-uint16_t namcos21_state::read_word_from_slave_input()
-{
-	uint16_t data = 0;
-	if( m_mpDspState->slaveBytesAvailable>0 )
-	{
-		data = m_mpDspState->slaveInputBuffer[m_mpDspState->slaveInputStart++];
-		m_mpDspState->slaveInputStart %= DSP_BUF_MAX;
-		m_mpDspState->slaveBytesAvailable--;
-		if( m_mpDspState->slaveBytesAdvertised>0 )
-		{
-			m_mpDspState->slaveBytesAdvertised--;
-		}
-		if (ENABLE_LOGGING) logerror( "%s:-%04x(0x%04x)\n", machine().describe_context(), data, m_mpDspState->slaveBytesAvailable );
-	}
-	return data;
-}
-
-uint16_t namcos21_state::get_input_bytes_advertised_for_slave()
-{
-	if( m_mpDspState->slaveBytesAdvertised < m_mpDspState->slaveBytesAvailable )
-	{
-		m_mpDspState->slaveBytesAdvertised++;
-	}
-	else if( m_mpDspState->slaveActive && m_mpDspState->masterFinished && m_mpDspState->masterSourceAddr )
-	{
-		namcos21_kickstart_hacks(0);
-	}
-	return m_mpDspState->slaveBytesAdvertised;
-}
-
-READ16_MEMBER(namcos21_state::dspram16_r)
-{
-	return m_dspram16[offset];
-}
-
-template<bool maincpu> WRITE16_MEMBER(namcos21_state::dspram16_w)
-{
-	COMBINE_DATA(&m_dspram16[offset]);
-
-	if (m_mpDspState->masterSourceAddr && offset == 1 + (m_mpDspState->masterSourceAddr & 0x7fff))
-	{
-		if (ENABLE_LOGGING) logerror("IDC-CONTINUE\n");
-		transfer_dsp_data();
-	}
-	else if (m_gametype == NAMCOS21_SOLVALOU && offset == 0x103 && maincpu)
-	{
-		/* hack; synchronization for solvalou */
-		m_maincpu->yield();
-	}
-}
-
-
-/************************************************************************************/
-
-
-/***********************************************************/
-
-
-
-READ16_MEMBER(namcos21_state::dsp_port0_r)
-{
-	int32_t data = read_pointrom_data(m_pointrom_idx++);
-	m_mPointRomMSB = (uint8_t)(data>>16);
-	m_mbPointRomDataAvailable = 1;
-	return (uint16_t)data;
-}
-
-WRITE16_MEMBER(namcos21_state::dsp_port0_w)
-{ /* unused? */
-	if (ENABLE_LOGGING) logerror( "PTRAM_LO(0x%04x)\n", data );
-}
-
-READ16_MEMBER(namcos21_state::dsp_port1_r)
-{
-	if( m_mbPointRomDataAvailable )
-	{
-		m_mbPointRomDataAvailable = 0;
-		return m_mPointRomMSB;
-	}
-	return 0x8000; /* IDC ack? */
-}
-
-WRITE16_MEMBER(namcos21_state::dsp_port1_w)
-{ /* unused? */
-	if (ENABLE_LOGGING) logerror( "PTRAM_HI(0x%04x)\n", data );
-}
-
-READ16_MEMBER(namcos21_state::dsp_port2_r)
-{ /* IDC TRANSMIT ENABLE? */
-	return 0;
-}
-
-WRITE16_MEMBER(namcos21_state::dsp_port2_w)
-{
-	if (ENABLE_LOGGING) logerror( "IDC ADDR INIT(0x%04x)\n", data );
-	m_mpDspState->masterSourceAddr = data;
-	transfer_dsp_data();
-}
-
-READ16_MEMBER(namcos21_state::dsp_port3_idc_rcv_enable_r)
-{ /* IDC RECEIVE ENABLE? */
-	return 0;
-}
-
-WRITE16_MEMBER(namcos21_state::dsp_port3_w)
-{
-	m_pointrom_idx<<=16;
-	m_pointrom_idx|=data;
-}
-
-WRITE16_MEMBER(namcos21_state::dsp_port4_w)
-{ /* receives $0B<<4 prior to IDC setup */
-}
-
-READ16_MEMBER(namcos21_state::dsp_port8_r)
-{ /* SMU status */
-	return 1;
-}
-
-
-WRITE16_MEMBER(namcos21_state::dsp_port8_w)
-{
-	if (ENABLE_LOGGING) logerror( "port8_w(%d)\n", data );
-	if( data )
-	{
-		m_mpDspState->masterFinished = 1;
-	}
-	m_irq_enable = data;
-}
-
-READ16_MEMBER(namcos21_state::dsp_port9_r)
-{ /* render-device-busy; used for direct-draw */
-	return 0;
-}
-
-READ16_MEMBER(namcos21_state::dsp_porta_r)
-{ /* config */
-	return 0;
-}
-
-WRITE16_MEMBER(namcos21_state::dsp_porta_w)
-{
-	/* boot: 1 */
-	/* IRQ0 end: 0 */
-	/* INT2 begin: 1 */
-	/* direct-draw begin: 0 */
-	/* INT1 begin: 1 */
-//  if (ENABLE_LOGGING) logerror( "dsp_porta_w(0x%04x)\n", data );
-}
-
-READ16_MEMBER(namcos21_state::dsp_portb_r)
-{ /* config */
-	return 1;
-}
-
-WRITE16_MEMBER(namcos21_state::dsp_portb_w)
-{
-	if( data==0 )
-	{ /* only 0->1 transition triggers */
-		return;
-	}
-	if( m_mpDspState->masterDirectDrawSize == 13 )
-	{
-		int i;
-		int sx[4], sy[4], zcode[4];
-		int color  = m_mpDspState->masterDirectDrawBuffer[0];
-		for( i=0; i<4; i++ )
-		{
-			sx[i] = NAMCOS21_POLY_FRAME_WIDTH/2 + (int16_t)m_mpDspState->masterDirectDrawBuffer[i*3+1];
-			sy[i] = NAMCOS21_POLY_FRAME_HEIGHT/2 + (int16_t)m_mpDspState->masterDirectDrawBuffer[i*3+2];
-			zcode[i] = m_mpDspState->masterDirectDrawBuffer[i*3+3];
-		}
-		if( color&0x8000 )
-		{
-			m_namcos21_3d->draw_quad(sx, sy, zcode, color);
-		}
-		else
-		{
-			logerror( "indirection used w/ direct draw?\n" );
-		}
-	}
-	else if( m_mpDspState->masterDirectDrawSize )
-	{
-		logerror( "unexpected masterDirectDrawSize=%d!\n",m_mpDspState->masterDirectDrawSize );
-	}
-	m_mpDspState->masterDirectDrawSize = 0;
-}
-
-WRITE16_MEMBER(namcos21_state::dsp_portc_w)
-{
-	if( m_mpDspState->masterDirectDrawSize < DSP_BUF_MAX )
-	{
-		m_mpDspState->masterDirectDrawBuffer[m_mpDspState->masterDirectDrawSize++] = data;
-	}
-	else
-	{
-		logerror( "portc overflow\n" );
-	}
-}
-
-READ16_MEMBER(namcos21_state::dsp_portf_r)
-{ /* informs BIOS that this is Master DSP */
-	return 0;
-}
-
-WRITE16_MEMBER(namcos21_state::dsp_xf_w)
-{
-	if (ENABLE_LOGGING) logerror("xf(%d)\n",data);
-}
-
-void namcos21_state::master_dsp_program(address_map &map)
-{
-	map(0x8000, 0xbfff).ram().share("master_dsp_ram");
-}
-
-void namcos21_state::master_dsp_data(address_map &map)
-{
-	map(0x2000, 0x200f).rw(FUNC(namcos21_state::dspcuskey_r), FUNC(namcos21_state::dspcuskey_w));
-	map(0x8000, 0xffff).rw(FUNC(namcos21_state::dspram16_r), FUNC(namcos21_state::dspram16_w<false>)); /* 0x8000 words */
-}
-
-void namcos21_state::master_dsp_io(address_map &map)
-{
-	map(0x00, 0x00).rw(FUNC(namcos21_state::dsp_port0_r), FUNC(namcos21_state::dsp_port0_w));
-	map(0x01, 0x01).rw(FUNC(namcos21_state::dsp_port1_r), FUNC(namcos21_state::dsp_port1_w));
-	map(0x02, 0x02).rw(FUNC(namcos21_state::dsp_port2_r), FUNC(namcos21_state::dsp_port2_w));
-	map(0x03, 0x03).rw(FUNC(namcos21_state::dsp_port3_idc_rcv_enable_r), FUNC(namcos21_state::dsp_port3_w));
-	map(0x04, 0x04).w(FUNC(namcos21_state::dsp_port4_w));
-	map(0x08, 0x08).rw(FUNC(namcos21_state::dsp_port8_r), FUNC(namcos21_state::dsp_port8_w));
-	map(0x09, 0x09).r(FUNC(namcos21_state::dsp_port9_r));
-	map(0x0a, 0x0a).rw(FUNC(namcos21_state::dsp_porta_r), FUNC(namcos21_state::dsp_porta_w));
-	map(0x0b, 0x0b).rw(FUNC(namcos21_state::dsp_portb_r), FUNC(namcos21_state::dsp_portb_w));
-	map(0x0c, 0x0c).w(FUNC(namcos21_state::dsp_portc_w));
-	map(0x0f, 0x0f).r(FUNC(namcos21_state::dsp_portf_r));
-}
-
-/************************************************************************************/
-
-void namcos21_state::render_slave_output(uint16_t data)
-{
-	if( m_mpDspState->slaveOutputSize >= 4096 )
-	{
-		fatalerror( "SLAVE OVERFLOW (0x%x)\n",m_mpDspState->slaveOutputBuffer[0]  );
-	}
-
-	/* append word to slave output buffer */
-	m_mpDspState->slaveOutputBuffer[m_mpDspState->slaveOutputSize++] = data;
-
-	{
-		uint16_t *pSource = m_mpDspState->slaveOutputBuffer;
-		uint16_t count = *pSource++;
-		if( count && m_mpDspState->slaveOutputSize > count )
-		{
-			uint16_t color = *pSource++;
-			int sx[4], sy[4],zcode[4];
-			int j;
-			if( color&0x8000 )
-			{
-				if( count!=13 ) logerror( "?!direct-draw(%d)\n", count );
-				for( j=0; j<4; j++ )
-				{
-					sx[j] = NAMCOS21_POLY_FRAME_WIDTH/2 + (int16_t)pSource[3*j+0];
-					sy[j] = NAMCOS21_POLY_FRAME_HEIGHT/2 + (int16_t)pSource[3*j+1];
-					zcode[j] = pSource[3*j+2];
-				}
-				m_namcos21_3d->draw_quad(sx, sy, zcode, color&0x7fff);
-			}
-			else
-			{
-				int quad_idx = color*6;
-				for(;;)
-				{
-					uint8_t code = m_pointram[quad_idx++];
-					color = m_pointram[quad_idx++]|(code<<8);
-					for( j=0; j<4; j++ )
-					{
-						uint8_t vi = m_pointram[quad_idx++];
-						sx[j] = NAMCOS21_POLY_FRAME_WIDTH/2  + (int16_t)pSource[vi*3+0];
-						sy[j] = NAMCOS21_POLY_FRAME_HEIGHT/2 + (int16_t)pSource[vi*3+1];
-						zcode[j] = pSource[vi*3+2];
-					}
-					m_namcos21_3d->draw_quad(sx, sy, zcode, color&0x7fff);
-					if( code&0x80 )
-					{ /* end-of-quadlist marker */
-						break;
-					}
-				}
-			}
-			m_mpDspState->slaveOutputSize = 0;
-		}
-		else if( count==0 )
-		{
-			fatalerror( "RenderSlaveOutput\n" );
-		}
-	}
-}
-
-READ16_MEMBER(namcos21_state::slave_port0_r)
-{
-	return read_word_from_slave_input();
-}
-
-WRITE16_MEMBER(namcos21_state::slave_port0_w)
-{
-	render_slave_output(data);
-}
-
-READ16_MEMBER(namcos21_state::slave_port2_r)
-{
-	return get_input_bytes_advertised_for_slave();
-}
-
-READ16_MEMBER(namcos21_state::slave_port3_r)
-{ /* render-device queue size */
-	/* up to 0x1fe bytes?
-	 * slave blocks until free &space exists
-	 */
-	return 0;
-}
-
-WRITE16_MEMBER(namcos21_state::slave_port3_w)
-{ /* 0=busy, 1=ready? */
-}
-
-WRITE16_MEMBER(namcos21_state::slave_XF_output_w)
-{
-	if (ENABLE_LOGGING) logerror( "%s :slaveXF(%d)\n", machine().describe_context(), data );
-}
-
-READ16_MEMBER(namcos21_state::slave_portf_r)
-{ /* informs BIOS that this is Slave DSP */
-	return 1;
-}
-
-void namcos21_state::slave_dsp_program(address_map &map)
-{
-	map(0x8000, 0x8fff).ram();
-}
-
-void namcos21_state::slave_dsp_data(address_map &map)
-{
-	/* no external data memory */
-}
-
-void namcos21_state::slave_dsp_io(address_map &map)
-{
-	map(0x00, 0x00).rw(FUNC(namcos21_state::slave_port0_r), FUNC(namcos21_state::slave_port0_w));
-	map(0x02, 0x02).r(FUNC(namcos21_state::slave_port2_r));
-	map(0x03, 0x03).rw(FUNC(namcos21_state::slave_port3_r), FUNC(namcos21_state::slave_port3_w));
-	map(0x0f, 0x0f).r(FUNC(namcos21_state::slave_portf_r));
-}
-
-/************************************************************************************/
-
-/**
- * 801f->800f : prepare for master access to point ram
- * 801f       : done
- *
- * #bits  data  line
- *   8     1a0    4
- *   7     0f8    4
- *   7     0ff    4
- *   1     001    4
- *   7     00a    2
- *   a     0fe    8
- *
- * line   #bits  data
- * 0003   000A   000004FE
- * 0001   0007   0000000A
- * 0002   001A   03FFF1A0
- */
-WRITE16_MEMBER(namcos21_state::pointram_control_w)
-{
-//  uint16_t prev = m_pointram_control;
-	COMBINE_DATA( &m_pointram_control );
-
-	/* m_pointram_control&0x20 : bank for depthcue data */
-#if 0
-	logerror( "%s dsp_control_w:[%x]:=%04x ",
-		machine().describe_context(),
-		offset,
-		m_pointram_control );
-
-	uint16_t delta = (prev^m_pointram_control)&m_pointram_control;
-	if( delta&0x10 )
-	{
-		logerror( " [reset]" );
-	}
-	if( delta&2 )
-	{
-		logerror( " send(A)%x", m_pointram_control&1 );
-	}
-	if( delta&4 )
-	{
-		logerror( " send(B)%x", m_pointram_control&1 );
-	}
-	if( delta&8 )
-	{
-		logerror( " send(C)%x", m_pointram_control&1 );
-	}
-	logerror( "\n" );
-#endif
-	m_pointram_idx = 0; /* HACK */
-}
-
-READ16_MEMBER(namcos21_state::pointram_data_r)
-{
-	return m_pointram[m_pointram_idx];
-}
-
-WRITE16_MEMBER(namcos21_state::pointram_data_w)
-{
-	if( ACCESSING_BITS_0_7 )
-	{
-//      if( (m_pointram_idx%6)==0 ) logerror("\n" );
-//      logerror( " %02x", data );
-		m_pointram[m_pointram_idx++] = data;
-		m_pointram_idx &= (PTRAM_SIZE-1);
-	}
-}
-
-
-READ16_MEMBER(namcos21_state::namcos21_depthcue_r)
-{
-	int bank = (m_pointram_control&0x20)?1:0;
-	return m_depthcue[bank][offset];
-}
-WRITE16_MEMBER(namcos21_state::namcos21_depthcue_w)
-{
-	if( ACCESSING_BITS_0_7 )
-	{
-		int bank = (m_pointram_control&0x20)?1:0;
-		m_depthcue[bank][offset] = data;
-//      if( (offset&0xf)==0 ) logerror( "\n depthcue: " );
-//      logerror( " %02x", data );
-	}
-}
 
 /* dual port ram memory handlers */
 
@@ -1219,10 +590,10 @@ WRITE8_MEMBER(namcos21_state::namcos2_dualportram_byte_w)
 void namcos21_state::common_map(address_map &map)
 {
 	map(0x280000, 0x280001).nopw(); /* written once on startup */
-	map(0x400000, 0x400001).w(FUNC(namcos21_state::pointram_control_w));
-	map(0x440000, 0x440001).rw(FUNC(namcos21_state::pointram_data_r), FUNC(namcos21_state::pointram_data_w));
+	map(0x400000, 0x400001).w(m_namcos21_dsp_c67, FUNC(namcos21_dsp_c67_device::pointram_control_w));
+	map(0x440000, 0x440001).rw(m_namcos21_dsp_c67, FUNC(namcos21_dsp_c67_device::pointram_data_r), FUNC(namcos21_dsp_c67_device::pointram_data_w));
 	map(0x440002, 0x47ffff).nopw(); /* (?) Air Combat */
-	map(0x480000, 0x4807ff).rw(FUNC(namcos21_state::namcos21_depthcue_r), FUNC(namcos21_state::namcos21_depthcue_w)); /* Air Combat */
+	map(0x480000, 0x4807ff).rw(m_namcos21_dsp_c67, FUNC(namcos21_dsp_c67_device::namcos21_depthcue_r), FUNC(namcos21_dsp_c67_device::namcos21_depthcue_w)); /* Air Combat */
 	map(0x700000, 0x71ffff).rw(m_c355spr, FUNC(namco_c355spr_device::spriteram_r), FUNC(namco_c355spr_device::spriteram_w));
 	map(0x720000, 0x720007).rw(m_c355spr, FUNC(namco_c355spr_device::position_r), FUNC(namco_c355spr_device::position_w));
 	map(0x740000, 0x74ffff).ram().w(m_palette, FUNC(palette_device::write16)).share("palette");
@@ -1243,7 +614,7 @@ void namcos21_state::master_map(address_map &map)
 	map(0x100000, 0x10ffff).ram(); /* private work RAM */
 	map(0x180000, 0x183fff).rw(FUNC(namcos21_state::namcos2_68k_eeprom_r), FUNC(namcos21_state::namcos2_68k_eeprom_w)).umask16(0x00ff);
 	map(0x1c0000, 0x1fffff).m(m_master_intc, FUNC(namco_c148_device::map));
-	map(0x200000, 0x20ffff).rw(FUNC(namcos21_state::dspram16_r), FUNC(namcos21_state::dspram16_w<true>)).share("dspram16");
+	map(0x200000, 0x20ffff).rw(m_namcos21_dsp_c67, FUNC(namcos21_dsp_c67_device::dspram16_r), FUNC(namcos21_dsp_c67_device::dspram16_hack_w)).share("namcos21dsp_c67:dspram16");
 }
 
 void namcos21_state::slave_map(address_map &map)
@@ -1252,7 +623,7 @@ void namcos21_state::slave_map(address_map &map)
 	map(0x000000, 0x07ffff).rom();
 	map(0x100000, 0x13ffff).ram(); /* private work RAM */
 	map(0x1c0000, 0x1fffff).m(m_slave_intc, FUNC(namco_c148_device::map));
-	map(0x200000, 0x20ffff).rw(FUNC(namcos21_state::dspram16_r), FUNC(namcos21_state::dspram16_w<false>)).share("dspram16");
+	map(0x200000, 0x20ffff).rw(m_namcos21_dsp_c67, FUNC(namcos21_dsp_c67_device::dspram16_r), FUNC(namcos21_dsp_c67_device::dspram16_w)).share("namcos21dsp_c67:dspram16");
 }
 
 void namcos21_state::winrun_master_map(address_map &map)
@@ -1263,8 +634,8 @@ void namcos21_state::winrun_master_map(address_map &map)
 	map(0x1c0000, 0x1fffff).m(m_master_intc, FUNC(namco_c148_device::map));
 	
 	// DSP Related
-	map(0x250000, 0x25ffff).ram().share("namcos21dsp:winrun_polydata"); // TOADD
-	map(0x260000, 0x26ffff).ram(); /* unused? */ //TOADD
+	map(0x250000, 0x25ffff).ram().share("namcos21dsp:winrun_polydata");
+	map(0x260000, 0x26ffff).ram(); /* unused? */
 	map(0x280000, 0x281fff).w(m_namcos21_dsp, FUNC(namcos21_dsp_device::winrun_dspbios_w)).share("namcos21dsp:winrun_dspbios");
 	map(0x380000, 0x38000f).rw(m_namcos21_dsp, FUNC(namcos21_dsp_device::winrun_dspcomram_control_r), FUNC(namcos21_dsp_device::winrun_dspcomram_control_w));
 	map(0x3c0000, 0x3c1fff).rw(m_namcos21_dsp, FUNC(namcos21_dsp_device::winrun_68k_dspcomram_r), FUNC(namcos21_dsp_device::winrun_68k_dspcomram_w));
@@ -1408,7 +779,7 @@ void namcos21_state::driveyes_master_map(address_map &map)
 	map(0x1c0000, 0x1fffff).m(m_master_intc, FUNC(namco_c148_device::map));
 	
 	// DSP related
-	map(0x250000, 0x25ffff).ram().share("namcos21dsp:winrun_polydata"); // TOADD
+	map(0x250000, 0x25ffff).ram().share("namcos21dsp:winrun_polydata");
 	map(0x280000, 0x281fff).w(m_namcos21_dsp, FUNC(namcos21_dsp_device::winrun_dspbios_w)).share("namcos21dsp:winrun_dspbios");
 	map(0x380000, 0x38000f).rw(m_namcos21_dsp, FUNC(namcos21_dsp_device::winrun_dspcomram_control_r), FUNC(namcos21_dsp_device::winrun_dspcomram_control_w));
 	map(0x3c0000, 0x3c1fff).rw(m_namcos21_dsp, FUNC(namcos21_dsp_device::winrun_68k_dspcomram_r), FUNC(namcos21_dsp_device::winrun_68k_dspcomram_w));
@@ -1710,13 +1081,10 @@ WRITE8_MEMBER(namcos21_state::sound_reset_w)
 		m_audiocpu->set_input_line(INPUT_LINE_RESET, ASSERT_LINE);
 	}
 
-	if (namcos2_kickstart)
+	if (data & 0x04)
 	{
-		//printf( "dspkick=0x%x\n", data );
-		if (data & 0x04)
-		{
-			namcos21_kickstart_hacks(1);
-		}
+		if (m_namcos21_dsp_c67)
+			m_namcos21_dsp_c67->reset_kickstart();
 	}
 }
 
@@ -1744,11 +1112,8 @@ void namcos21_state::reset_all_subcpus(int state)
 		logerror("no MCU to reset?\n");
 	}
 
-	if (m_c67master)
-		m_c67master->set_input_line(INPUT_LINE_RESET, state);
-
-	if (m_c67slave[0])
-		m_c67slave[0]->set_input_line(INPUT_LINE_RESET, state);
+	if (m_namcos21_dsp_c67)
+		m_namcos21_dsp_c67->reset_dsps(state);
 }
 
 
@@ -1775,17 +1140,6 @@ MACHINE_RESET_MEMBER(namcos21_state, namcos21)
 
 	/* Place CPU2 & CPU3 into the reset condition */
 	reset_all_subcpus(ASSERT_LINE);
-
-	/* DSP startup hacks */
-	m_mbNeedsKickstart = 20;
-	if( m_gametype==NAMCOS21_CYBERSLED )
-	{
-		m_mbNeedsKickstart = 200;
-	}
-
-	/* Wipe the framebuffers */
-	m_namcos21_3d->swap_and_clear_poly_framebuffer();
-	m_namcos21_3d->swap_and_clear_poly_framebuffer();
 }
 
 
@@ -1803,8 +1157,6 @@ MACHINE_START_MEMBER(namcos21_state,namcos21)
 
 		m_audiobank->set_entry(0);
 	}
-
-	namcos2_kickstart = true;
 }
 
 TIMER_DEVICE_CALLBACK_MEMBER(namcos21_state::screen_scanline)
@@ -1973,7 +1325,7 @@ MACHINE_CONFIG_START(namcos21_state::driveyes)
 	MCFG_SOUND_ROUTE(1, "rspeaker", 0.30)
 MACHINE_CONFIG_END
 
-// starblad, solvalou, aircomb, cybsled
+// starblad, solvalou, aircomb, cybsled base state
 MACHINE_CONFIG_START(namcos21_state::namcos21)
 	MCFG_DEVICE_ADD("maincpu", M68000,12288000) /* Master */
 	MCFG_DEVICE_PROGRAM_MAP(master_map)
@@ -1989,30 +1341,8 @@ MACHINE_CONFIG_START(namcos21_state::namcos21)
 
 	configure_c68_namcos21(config);
 
-	namco_c67_device& dspmaster(NAMCO_C67(config, m_c67master, 24000000)); /* 24 MHz? overclocked */
-	dspmaster.set_addrmap(AS_PROGRAM, &namcos21_state::master_dsp_program);
-	dspmaster.set_addrmap(AS_DATA, &namcos21_state::master_dsp_data);
-	dspmaster.set_addrmap(AS_IO, &namcos21_state::master_dsp_io);
-	dspmaster.hold_in_cb().set_constant(0);
-	dspmaster.hold_ack_out_cb().set_nop();
-	dspmaster.xf_out_cb().set(FUNC(namcos21_state::dsp_xf_w));
-
-	for (int i = 0; i < 4; i++)
-	{
-		namco_c67_device& dspslave(NAMCO_C67(config, m_c67slave[i], 24000000)); /* 24 MHz? overclocked */
-		dspslave.set_addrmap(AS_PROGRAM, &namcos21_state::slave_dsp_program);
-		dspslave.set_addrmap(AS_DATA, &namcos21_state::slave_dsp_data);
-		dspslave.set_addrmap(AS_IO, &namcos21_state::slave_dsp_io);
-		dspslave.hold_in_cb().set_constant(0);
-		dspslave.hold_ack_out_cb().set_nop();
-		dspslave.xf_out_cb().set(FUNC(namcos21_state::slave_XF_output_w));
-
-		// the emulation currently only uses one slave DSP clocked at 4x the normal rate instead of the master splitting the workload across the 4 slaves
-		if (i!=0)
-			dspslave.set_disable();
-		else
-			dspslave.set_clock(24000000*4);
-	}
+	NAMCOS21_DSP_C67(config, m_namcos21_dsp_c67, 0);
+	m_namcos21_dsp_c67->set_renderer_tag("namcos21_3d");
 
 	MCFG_QUANTUM_TIME(attotime::from_hz(12000))
 
@@ -2059,7 +1389,25 @@ MACHINE_CONFIG_START(namcos21_state::namcos21)
 	MCFG_SOUND_ROUTE(1, "rspeaker", 0.30)
 MACHINE_CONFIG_END
 
+MACHINE_CONFIG_START(namcos21_state::aircomb)
+	namcos21(config);
+	m_namcos21_dsp_c67->set_gametype(namcos21_dsp_c67_device::NAMCOS21_AIRCOMBAT);
+MACHINE_CONFIG_END
 
+MACHINE_CONFIG_START(namcos21_state::starblad)
+	namcos21(config);
+	m_namcos21_dsp_c67->set_gametype(namcos21_dsp_c67_device::NAMCOS21_STARBLADE);
+MACHINE_CONFIG_END
+
+MACHINE_CONFIG_START(namcos21_state::cybsled)
+	namcos21(config);
+	m_namcos21_dsp_c67->set_gametype(namcos21_dsp_c67_device::NAMCOS21_CYBERSLED);
+MACHINE_CONFIG_END
+
+MACHINE_CONFIG_START(namcos21_state::solvalou)
+	namcos21(config);
+	m_namcos21_dsp_c67->set_gametype(namcos21_dsp_c67_device::NAMCOS21_SOLVALOU);
+MACHINE_CONFIG_END
 
 
 
@@ -2315,7 +1663,7 @@ ROM_START( starblad )
 
 	ROM_REGION16_BE( 0x100000, "edata", ROMREGION_ERASEFF )
 
-	ROM_REGION32_BE( 0x400000, "point24", ROMREGION_ERASE00) /* 24bit signed point data */
+	ROM_REGION32_BE( 0x400000, "namcos21dsp_c67:point24", ROMREGION_ERASE00) /* 24bit signed point data */
 	ROM_LOAD32_BYTE( "st1-pt0-h.bin", 0x000001, 0x80000, CRC(84eb355f) SHA1(89a248b8be2e0afcee29ba4c4c9cca65d5fb246a) )
 	ROM_LOAD32_BYTE( "st1-pt0-u.bin", 0x000002, 0x80000, CRC(1956cd0a) SHA1(7d21b3a59f742694de472c545a1f30c3d92e3390) )
 	ROM_LOAD32_BYTE( "st1-pt0-l.bin", 0x000003, 0x80000, CRC(ff577049) SHA1(1e1595174094e88d5788753d05ce296c1f7eca75) )
@@ -2360,7 +1708,7 @@ ROM_START( starbladj )
 
 	ROM_REGION16_BE( 0x100000, "edata", ROMREGION_ERASEFF )
 
-	ROM_REGION32_BE( 0x400000, "point24", ROMREGION_ERASE00) /* 24bit signed point data */
+	ROM_REGION32_BE( 0x400000, "namcos21dsp_c67:point24", ROMREGION_ERASE00) /* 24bit signed point data */
 	ROM_LOAD32_BYTE( "st1-pt0-h.bin", 0x000001, 0x80000, CRC(84eb355f) SHA1(89a248b8be2e0afcee29ba4c4c9cca65d5fb246a) )
 	ROM_LOAD32_BYTE( "st1-pt0-u.bin", 0x000002, 0x80000, CRC(1956cd0a) SHA1(7d21b3a59f742694de472c545a1f30c3d92e3390) )
 	ROM_LOAD32_BYTE( "st1-pt0-l.bin", 0x000003, 0x80000, CRC(ff577049) SHA1(1e1595174094e88d5788753d05ce296c1f7eca75) )
@@ -2409,7 +1757,7 @@ ROM_START( solvalou )
 
 	ROM_REGION16_BE( 0x100000, "edata", ROMREGION_ERASEFF )
 
-	ROM_REGION32_BE( 0x400000, "point24", ROMREGION_ERASE00)       /* 24bit signed point data */
+	ROM_REGION32_BE( 0x400000, "namcos21dsp_c67:point24", ROMREGION_ERASE00)       /* 24bit signed point data */
 	ROM_LOAD32_BYTE( "sv1-pt0-h.bin", 0x000001, 0x80000, CRC(3be21115) SHA1(c9f30353c1216f64199f87cd34e787efd728e739) ) /* most significant */
 	ROM_LOAD32_BYTE( "sv1-pt0-u.bin", 0x000002, 0x80000, CRC(4aacfc42) SHA1(f0e179e057183b41744ca429764f44306f0ce9bf) )
 	ROM_LOAD32_BYTE( "sv1-pt0-l.bin", 0x000003, 0x80000, CRC(6a4dddff) SHA1(9ed182d21d328c6a684ee6658a9dfcf3f3dd8646) ) /* least significant */
@@ -2455,7 +1803,7 @@ ROM_START( aircomb )
 	ROM_LOAD16_BYTE( "ac1-edata1-u.3c", 0x000000, 0x80000, CRC(a9547509) SHA1(1bc663cec03b60ad968896bbc2546f02efda135e) )
 	ROM_LOAD16_BYTE( "ac1-edata1-l.1c", 0x000001, 0x80000, CRC(a87087dd) SHA1(cd9b83a8f07886ab44e4ded68002b44338777e8c) )
 
-	ROM_REGION32_BE( 0x400000, "point24", ROMREGION_ERASE00)       /* 24bit signed point data */
+	ROM_REGION32_BE( 0x400000, "namcos21dsp_c67:point24", ROMREGION_ERASE00)       /* 24bit signed point data */
 	ROM_LOAD32_BYTE( "ac1-poi-h.2f",  0x000001, 0x80000, CRC(573bbc3b) SHA1(371be12b915db6872049f18980c1b55544cfc445) ) /* most significant */
 	ROM_LOAD32_BYTE( "ac1-poi-lu.2k", 0x000002, 0x80000, CRC(d99084b9) SHA1(c604d60a2162af7610e5ff7c1aa4195f7df82efe) )
 	ROM_LOAD32_BYTE( "ac1-poi-ll.2n", 0x000003, 0x80000, CRC(abb32307) SHA1(8e936ba99479215dd33a951d81ec2b04020dfd62) ) /* least significant */
@@ -2508,7 +1856,7 @@ ROM_START( aircombj )
 	ROM_LOAD16_BYTE( "ac1-edata1-u.3c", 0x000000, 0x80000, CRC(a9547509) SHA1(1bc663cec03b60ad968896bbc2546f02efda135e) )
 	ROM_LOAD16_BYTE( "ac1-edata1-l.1c", 0x000001, 0x80000, CRC(a87087dd) SHA1(cd9b83a8f07886ab44e4ded68002b44338777e8c) )
 
-	ROM_REGION32_BE( 0x400000, "point24", ROMREGION_ERASE00)       /* 24bit signed point data */
+	ROM_REGION32_BE( 0x400000, "namcos21dsp_c67:point24", ROMREGION_ERASE00)       /* 24bit signed point data */
 	ROM_LOAD32_BYTE( "ac1-poi-h.2f",  0x000001, 0x80000, CRC(573bbc3b) SHA1(371be12b915db6872049f18980c1b55544cfc445) ) /* most significant */
 	ROM_LOAD32_BYTE( "ac1-poi-lu.2k", 0x000002, 0x80000, CRC(d99084b9) SHA1(c604d60a2162af7610e5ff7c1aa4195f7df82efe) )
 	ROM_LOAD32_BYTE( "ac1-poi-ll.2n", 0x000003, 0x80000, CRC(abb32307) SHA1(8e936ba99479215dd33a951d81ec2b04020dfd62) ) /* least significant */
@@ -2561,7 +1909,7 @@ ROM_START( cybsled )
 	ROM_LOAD16_BYTE( "cy1-edata0-u.3b", 0x000000, 0x80000, CRC(77452533) SHA1(48fc199bcc1beb23c714eebd9b09b153c980170b) )
 	ROM_LOAD16_BYTE( "cy1-edata0-l.1b", 0x000001, 0x80000, CRC(e812e290) SHA1(719e0a026ae8ef63d0d0269b67669ea9b4d950dd) )
 
-	ROM_REGION32_BE( 0x400000, "point24", ROMREGION_ERASE00)       /* 24bit signed point data */
+	ROM_REGION32_BE( 0x400000, "namcos21dsp_c67:point24", ROMREGION_ERASE00)       /* 24bit signed point data */
 	ROM_LOAD32_BYTE( "cy1-poi-h1.2f",  0x000001, 0x80000, CRC(eaf8bac3) SHA1(7a2caf6672af158b4a23ce4626342d1f17d1a4e4) )    /* most significant */
 	ROM_LOAD32_BYTE( "cy1-poi-lu1.2k", 0x000002, 0x80000, CRC(c544a8dc) SHA1(4cce5f2ab3519b4aa7edbdd15b2d79a7fdcade3c) )
 	ROM_LOAD32_BYTE( "cy1-poi-ll1.2n", 0x000003, 0x80000, CRC(30acb99b) SHA1(a28dcb3e5405f166644f6353a903c1143ee268f1) )    /* least significant */
@@ -2612,7 +1960,7 @@ ROM_START( cybsledj )
 	ROM_LOAD16_BYTE( "cy1-edata0-u.3b", 0x000000, 0x80000, CRC(77452533) SHA1(48fc199bcc1beb23c714eebd9b09b153c980170b) )
 	ROM_LOAD16_BYTE( "cy1-edata0-l.1b", 0x000001, 0x80000, CRC(e812e290) SHA1(719e0a026ae8ef63d0d0269b67669ea9b4d950dd) )
 
-	ROM_REGION32_BE( 0x400000, "point24", ROMREGION_ERASE00)       /* 24bit signed point data */
+	ROM_REGION32_BE( 0x400000, "namcos21dsp_c67:point24", ROMREGION_ERASE00)       /* 24bit signed point data */
 	ROM_LOAD32_BYTE( "cy1-poi-h1.2f",  0x000001, 0x80000, CRC(eaf8bac3) SHA1(7a2caf6672af158b4a23ce4626342d1f17d1a4e4) )    /* most significant */
 	ROM_LOAD32_BYTE( "cy1-poi-lu1.2k", 0x000002, 0x80000, CRC(c544a8dc) SHA1(4cce5f2ab3519b4aa7edbdd15b2d79a7fdcade3c) )
 	ROM_LOAD32_BYTE( "cy1-poi-ll1.2n", 0x000003, 0x80000, CRC(30acb99b) SHA1(a28dcb3e5405f166644f6353a903c1143ee268f1) )    /* least significant */
@@ -2630,31 +1978,6 @@ ROM_START( cybsledj )
 	ROM_LOAD( "cybsledj.nv", 0x0000, 0x2000, CRC(a73bb03e) SHA1(e074bfeae14178c867070e06f6690ed13115f5fa) )
 ROM_END
 
-
-void namcos21_state::init(int game_type)
-{
-	m_gametype = game_type;
-	m_pointram = std::make_unique<uint8_t[]>(PTRAM_SIZE);
-	m_mpDspState = make_unique_clear<dsp_state>();
-
-}
-
-void namcos21_state::init_aircomb()
-{
-	init(NAMCOS21_AIRCOMBAT);
-}
-
-void namcos21_state::init_starblad()
-{
-	init(NAMCOS21_STARBLADE);
-}
-
-
-void namcos21_state::init_cybsled()
-{
-	init(NAMCOS21_CYBERSLED);
-}
-
 void namcos21_state::init_solvalou()
 {
 	uint16_t *mem = (uint16_t *)memregion("maincpu")->base();
@@ -2662,21 +1985,12 @@ void namcos21_state::init_solvalou()
 	mem[0x20cf4/2+0] = 0x4e71; // 2nd ptr_booting
 	mem[0x20cf4/2+1] = 0x4e71;
 	mem[0x20cf4/2+2] = 0x4e71;
-
-	init(NAMCOS21_SOLVALOU );
-}
-
-void namcos21_state::init_driveyes()
-{
-	m_mbNeedsKickstart = 0;
 }
 
 void namcos21_state::init_winrun()
 {
-	m_videoram = std::make_unique<uint8_t[]>(0x80000);
-	m_maskram = std::make_unique<uint8_t[]>(0x80000);
-
-	m_mbNeedsKickstart = 0;
+	m_gpu_videoram = std::make_unique<uint8_t[]>(0x80000);
+	m_gpu_maskram = std::make_unique<uint8_t[]>(0x80000);
 }
 
 
@@ -2689,13 +2003,13 @@ GAME( 1989, winrungp,  0,        winrun,   winrungp,   namcos21_state, init_winr
 GAME( 1991, winrun91,  0,        winrun,   winrungp,   namcos21_state, init_winrun,   ROT0,    "Namco", "Winning Run '91 (Japan) (1991/03/05, Main Ver 1.0, Sub Ver 1.0)",               MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN )
 
 // 3 PCB stacks in a single cage (3x 4 PCBs) linked for 3 screen panorama, boards look similar to original Namco System 21 (not 21B) including TMS320C25 DSP, but use C68 I/O MCU and sprite chip instead of "68000 'GPU'" ?
-GAME( 1992, driveyes,  0,        driveyes, driveyes,   namcos21_state, init_driveyes, ROT0,    "Namco", "Driver's Eyes (Japan) (1992/01/10, Main Ver 2.1, Sub Ver 1.1)",                 MACHINE_NOT_WORKING | MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN)
+GAME( 1992, driveyes,  0,        driveyes, driveyes,   namcos21_state, empty_init,    ROT0,    "Namco", "Driver's Eyes (Japan) (1992/01/10, Main Ver 2.1, Sub Ver 1.1)",                 MACHINE_NOT_WORKING | MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN)
 
 // uses 5x TMS320C25 (C67, has internal ROM - dumped) 
-GAME( 1991, starblad,  0,        namcos21, starblad,   namcos21_state, init_starblad, ROT0,    "Namco", "Starblade (World)",                     MACHINE_IMPERFECT_GRAPHICS )
-GAME( 1991, starbladj, starblad, namcos21, starblad,   namcos21_state, init_starblad, ROT0,    "Namco", "Starblade (Japan)",                     MACHINE_IMPERFECT_GRAPHICS )
-GAME( 1991, solvalou,  0,        namcos21, s21default, namcos21_state, init_solvalou, ROT0,    "Namco", "Solvalou (Japan)",                      MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING )
-GAME( 1992, aircomb,   0,        namcos21, aircomb,    namcos21_state, init_aircomb,  ROT0,    "Namco", "Air Combat (US)",                       MACHINE_NOT_WORKING | MACHINE_IMPERFECT_GRAPHICS ) // There's code for a SCI, is it even possible to play multiplayer?
-GAME( 1992, aircombj,  aircomb,  namcos21, aircomb,    namcos21_state, init_aircomb,  ROT0,    "Namco", "Air Combat (Japan)",                    MACHINE_NOT_WORKING | MACHINE_IMPERFECT_GRAPHICS )
-GAME( 1993, cybsled,   0,        namcos21, cybsled,    namcos21_state, init_cybsled,  ROT0,    "Namco", "Cyber Sled (World)",                    MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN | MACHINE_NOT_WORKING )
-GAME( 1993, cybsledj,  cybsled,  namcos21, cybsled,    namcos21_state, init_cybsled,  ROT0,    "Namco", "Cyber Sled (Japan)",                    MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN | MACHINE_NOT_WORKING )
+GAME( 1991, starblad,  0,        starblad, starblad,   namcos21_state, empty_init,    ROT0,    "Namco", "Starblade (World)",                     MACHINE_IMPERFECT_GRAPHICS )
+GAME( 1991, starbladj, starblad, starblad, starblad,   namcos21_state, empty_init,    ROT0,    "Namco", "Starblade (Japan)",                     MACHINE_IMPERFECT_GRAPHICS )
+GAME( 1991, solvalou,  0,        solvalou, s21default, namcos21_state, init_solvalou, ROT0,    "Namco", "Solvalou (Japan)",                      MACHINE_IMPERFECT_GRAPHICS | MACHINE_NOT_WORKING )
+GAME( 1992, aircomb,   0,        aircomb,  aircomb,    namcos21_state, empty_init,    ROT0,    "Namco", "Air Combat (US)",                       MACHINE_NOT_WORKING | MACHINE_IMPERFECT_GRAPHICS ) // There's code for a SCI, is it even possible to play multiplayer?
+GAME( 1992, aircombj,  aircomb,  aircomb,  aircomb,    namcos21_state, empty_init,    ROT0,    "Namco", "Air Combat (Japan)",                    MACHINE_NOT_WORKING | MACHINE_IMPERFECT_GRAPHICS )
+GAME( 1993, cybsled,   0,        cybsled,  cybsled,    namcos21_state, empty_init,    ROT0,    "Namco", "Cyber Sled (World)",                    MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN | MACHINE_NOT_WORKING )
+GAME( 1993, cybsledj,  cybsled,  cybsled,  cybsled,    namcos21_state, empty_init,    ROT0,    "Namco", "Cyber Sled (Japan)",                    MACHINE_IMPERFECT_GRAPHICS | MACHINE_NODEVICE_LAN | MACHINE_NOT_WORKING )
