+++ src/mame/machine/namcos21_dsp.h
@@ -24,16 +24,6 @@ public:
 	// config
 	template <typename T> void set_renderer_tag(T &&tag) { m_renderer.set_tag(std::forward<T>(tag)); }
 
-	DECLARE_READ16_MEMBER(winrun_dspcomram_r);
-	DECLARE_WRITE16_MEMBER(winrun_dspcomram_w);
-	DECLARE_READ16_MEMBER(winrun_cuskey_r);
-	DECLARE_WRITE16_MEMBER(winrun_cuskey_w);
-	DECLARE_READ16_MEMBER(winrun_poly_reset_r);
-	DECLARE_WRITE16_MEMBER(winrun_dsp_render_w);
-	DECLARE_WRITE16_MEMBER(winrun_dsp_pointrom_addr_w);
-	DECLARE_READ16_MEMBER(winrun_dsp_pointrom_data_r);
-	DECLARE_WRITE16_MEMBER(winrun_dsp_complete_w);
-	DECLARE_READ16_MEMBER(winrun_table_r);
 	DECLARE_WRITE16_MEMBER(winrun_dspbios_w);
 	DECLARE_READ16_MEMBER(winrun_68k_dspcomram_r);
 	DECLARE_WRITE16_MEMBER(winrun_68k_dspcomram_w);
@@ -74,6 +64,18 @@ private:
 	int m_winrun_dsp_alive;
 
 	void winrun_flush_poly();
+
+	DECLARE_READ16_MEMBER(winrun_cuskey_r);
+	DECLARE_WRITE16_MEMBER(winrun_cuskey_w);
+	DECLARE_READ16_MEMBER(winrun_dspcomram_r);
+	DECLARE_WRITE16_MEMBER(winrun_dspcomram_w);
+	DECLARE_READ16_MEMBER(winrun_table_r);
+	DECLARE_WRITE16_MEMBER(winrun_dsp_complete_w);
+	DECLARE_WRITE16_MEMBER(winrun_dsp_render_w);
+	DECLARE_READ16_MEMBER(winrun_poly_reset_r);
+	DECLARE_WRITE16_MEMBER(winrun_dsp_pointrom_addr_w);
+	DECLARE_READ16_MEMBER(winrun_dsp_pointrom_data_r);
+
 };
 
 DECLARE_DEVICE_TYPE(NAMCOS21_DSP, namcos21_dsp_device)
